Move supported PEP 695 features to main annotations
