+++ tests/annotations_unsupported.pyi
@@ -1,13 +1,6 @@
 # Generated via: manual separation of unsupported features
 # These declarations use syntax from PEP 695 that mypy fails to parse.
-from typing import (
-    Callable,
-    NewType,
-    ParamSpec,
-    TypeVar,
-    TypeVarTuple,
-    Unpack,
-)
+from typing import NewType, ParamSpec, TypeVar, TypeVarTuple
 
 T = TypeVar("T")
 P = ParamSpec("P")
@@ -15,38 +8,6 @@ Ts = TypeVarTuple("Ts")
 U = TypeVar("U")
 NumberLike = TypeVar("NumberLike")
 UserId = NewType("UserId", int)
-
-MyList = list[int]
-
-type StrList = list[str]
-
-type Alias0[T] = list[T]
-
-type Alias1[T] = Alias0[T]
-
-type AliasNewType = UserId
-
-type AliasTypeVar[T] = T
-
-type AliasUnion = int | str
-
-type ListOrSet[T] = list[T] | set[T]
-
-type IntFunc[**P] = Callable[P, int]
-
-type LabeledTuple[*Ts] = tuple[str, Unpack[Ts]]
-
-type RecursiveList[T] = T | list[RecursiveList[T]]
-
-class NewGeneric[T]:
-    value: T
-    def get(self) -> T: ...
-
-class BoundClass[T: int]:
-    value: T
-
-class ConstrainedClass[T: (int, str)]:
-    value: T
 
 def use_params(*args: P.args, **kwargs: P.kwargs) -> int: ...
